requirements.txt
alembic==1.14.1
bcrypt==4.1.2
blinker==1.8.2
certifi==2025.1.31
click==8.1.8
distlib==0.3.9
exceptiongroup==1.2.2
Faker==18.11.2
filelock==3.16.1
Flask==2.0.3
Flask-SQLAlchemy==2.5.1
Flask-Migrate==3.1.0
Flask-JWT-Extended==4.4.4
Flask-Mail==0.9.1
Flask-Uploads==0.2.1
cloudinary==1.25.0
pytest==6.2.4
psycopg2-binary==2.9.1
gunicorn==20.1.0
greenlet==3.1.1
importlib_metadata==8.5.0
importlib_resources==6.4.5
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==2.1.5
marshmallow==3.19.0
packaging==25.0
pipenv==2024.4.1
platformdirs==4.3.6
pluggy==0.13.1
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
six==1.17.0
SQLAlchemy==2.0.40
tomli==2.2.1
typing_extensions==4.13.2
virtualenv==20.30.0
werkzeug==2.0.3
zipp==3.20.2

Flask-SQLAlchemy==2.5.1
Flask-Migrate==3.1.0
Flask-JWT-Extended==4.4.4
Flask-Mail==0.9.1

